Make all i18n tags compatible with the Zope 2.7 implementation.
The Zope 2.7 implementation is more restrictive that AltPTi18n, in that
(for internal implementation reasons) it can't have a tal:replace and a
i18n:translate on the same tag. This means that constructions like:

 <span tal:replace="foo/bar" i18n:translate="">Foobar</span>

have to be replaced by the more verbose:

 <span tal:omit-tag="" i18n:translate=""><span tal:replace="foo/bar">Foobar</span></span>

This basically corresponds to computed message ids. There are not many
such occurences in Plone so it's not a big problem to use the verbose
form.

svn path=/CMFPlone/trunk/; revision=1233