Im probably breaking discussion here but hey thats why we have cvs. So someone can back out all these changes tommorrow...
discussionitem_view does not need to do any testing on the view, that is all done in discussionView macro. For some reason if by changing the way this is called template.getId() in the discussionView is now read correctly. I need to read up more on how this works.

viewThreadsAtBottom needs to have the isDiscussionAllowed and isViewTemplate in a separate check, to allow the previous call as far as i can tell. Now template.getId() works and the isDiscussionAllowed and isViewTemplate check passes.

There are still things I dont like about the discussion interface, but i want to make sure im on the right track first...

svn path=/CMFPlone/trunk/; revision=1238