large fix for personalBar.
workflow cruft was being injected into the personalBar.
this cruft was not i18nable and it broke the semantics
of the personalBar.

getOrderedUserActions places the actions in a sane order
(configurable through parameters)

keyFilteredActions given filterdActionsFor dictioanry
will create a two dimensional dictionary [category][action_id]
this is *very* handy.  the two dimensonal dictionary is
also being called in global_defines.

if someone would look at the code to see if its "optimal"
that would be nice *wink*

svn path=/CMFPlone/branches/Plone-2_0-branch/; revision=3112