Making table-based columns the default again - and adding a plone_tableless for whose who want to do tableless designs.
For those of you who were running tableless designs in the RC3
and earlier releases and want to keep it that way - simply add
a DirectoryView for 'plone_tableless' in 'portal_skins', and
add it to the skinpath. This will activate the tableless versions
of:

- main_template

- ploneColumns.css

Apart from adding these two templates, no other changes are needed.

svn path=/CMFPlone/branches/Plone-2_0-branch/; revision=4163