search.pt can be a CMFTopic content's view out of box since  CMFTopic has its own queryCatalog method too:
    def queryCatalog( self, REQUEST=None, **kw ):

I think the search form is currently better than the topic_view. 

Just added the request parameter. This way, additional query conditions from 
the request can be used in CMFTopic's own queryCatalog.

We need such a feature. And it is a very minor fix, so I check it into svn. 

svn path=/CMFPlone/branches/2.1/; revision=5903