Added the sanitizeForURL script. This is a script to automate Short Names, anchors, CSS class name generation - anything that needs to generate a sensible ID.
Will have to be put in unrestricted Python code somewhere because of the re usage,
but I'll stick it here for the time being, so it won't get lost if my computer
crashes/is stolen (which has happened before ;). We are planning to work on this
on the UI sprint in Amsterdam.

Doctests included.

Made by Anders Pearson, from spec by Limi - thanks, Anders!

svn path=/CMFPlone/branches/2.1/; revision=6013