Merging PLIP75 - A blacklisting of types (unfriendly_types in site_properties) for searching etc. A method in PloneTool.py does the filtering, and is called from queryCatalog.py to affect the regular Plone searches. Migrations and tests and oh, my.
Also: A missing migration from the PLIP77-78 branch was caught and added,
giving us the non_default_page_types property in site_properties, and a slightly
more resilient select_default_page.cpt, which was barfing when this property
was missing.

NOTE: uiteam-plip75-search-blacklist is now dead and closed, and does *not*
contain all the changes in this commit, as the branch was too outdated to 
continue work on.

svn path=/CMFPlone/branches/2.1/; revision=6505