Commiting fix for collector issue 3704.
navigationParent.py has been cleaned up and refactored to be more generically
useful. It no longer hardcodes '/folder_contents' in the url when 
templateId == 'folder_contents'; that special case is handled by folder_contents
as it should be. Similarly, it doesn't hardcode a check for the
'List folder contents' permission at the parent - instead a generic list of
permissions can be passed in (and again, this is done by folder_contents,
which was the only page using it before).

By default, navigationParent "falls through" an index_html/default_page
object. If you are viewing an index_html page and want the parent, you
probably want the parent-of-parent, since going to the parent will land
you right back at the object you were viewing. 

The exception is when you want to call an explicit page template on the 
parent (such as folder_contents). In this case, a flag fallThroughDefaultPage 
can be set to False (folder_contents does this), and you will get the 
immediate parent.

The script is now also better at handling permissions (e.g. if you can't
view the parent object, it doesn't raise an exception, simply returns None), 
and can handle the case where you have a folderish index_html containing its 
own index_html (recursively ad infinitum, if need be). It should also be
a tad more efficient.

Any the tests are almost done running now, so I can stop typing and press
save. :)


svn path=/CMFPlone/branches/2.1/; revision=6589