merge in fix for #3093 thanks alec the charset issues are not going to go away ;-(
in fact this 'transaction' note stuff is just
plain wrong.. but the concept is sound.  should
be at the application level not in the zodb ;-)

NOTE: the audit log is the best place to put this
which we have working and should be released in
the next few weeks.

transaction notes/undo msgs are exposed in the
UI of Plone - most of this baggage is hysertical.


svn path=/CMFPlone/branches/2.1/; revision=6805