With Plone 2.1's reliance on the catalog metadata/index is_folderish to make decisions about rendering portal tabs, folder_contents linking to folder_contents and displaying the Contents tab, it becomes very difficult to use custom types which are folderish as an implementation detail but in general should not act as a folder in the interface. This totally broke RichDocument, for example, which in 2.0.5 was able to work around the issues with now-deprecated navtree and site properties.
Since setting isPrincipiaFolderish to False on such items has too many side 
effects, a better way of differentiating between "structural" folders and 
"non-structural folderish items" is needed. A simple marker interface has been 
added for this, which is checked in the catalog wrapper, with a helper method +
script and some tests. 

svn path=/CMFPlone/branches/2.1/; revision=7890