Reverted 8058 because
a) It is *really* late in the release cycle, and

b) Quoting the Catalog.search() docstring:
"""Limit is used in conjunction with sorting or scored results to inform
the catalog how many results you are really interested in. The catalog
can then use optimizations to save time and memory. The number of
results is not guaranteed to fall within the limit however, you should
still slice or batch the results as usual."""

c) No tests?


svn path=/CMFPlone/branches/2.1/; revision=8059