fixeS #5204: livesearch works with ++ and otherwise fixed too
Few issues fixed, some is extra to the original report:

- The javascript properly escapes + when making the XMLHttpRequest
- named entities are replaced by numbered ones in the response
   (this blathers error in some browsers since the response is xml not html)
- the "more items" query in the response is properly escaped too
- strip all meta characters from the query (+*-?)

Otherwise C++ does not really look for C++ but for C. The same
happens with other characters like *, -, ? too. Problem is the
catalog splits on words and all non-word chars are handled
equally like a space. We can do more once the catalog handles
these in another way.


svn path=/Plone3rdParty/branches/2.1/; revision=9785