Made the core content types use CSS sprites instead of individual images, which cuts down the number of image HTTP requests from 11 to 2. This fixes #10403.
Cleaned up the type descriptions, and removed the ones where that are 
self-evident or redundant.

Totally cargo-culted the plone.app.upgrade code, so please have a look to see 
if it makes sense. Some types have empty descriptions, this is intentional.

Also cleaned up the templates that have icons associated with them (listings,
portlets).


svn path=/plone.app.portlets/trunk/; revision=35801