Pass in the context when getting action information.
If you do not, then protected actions cannot determine the correct access level and you can get Unauthorized errors. This only affects content items that do not implement IBrowserDefault and for urls other than the default view or with urls ending in /view.

Writing a test for this would involve creating a non-ATContentTypes (non-dynamic view) content item, restricting it's view action and then testing view_template_id for unexpected Unauthorized errors. A lot of work for a 16 character fix.

svn path=/plone.app.layout/trunk/; revision=40133