Use portal_type for the "Item Type" ATPortalTypeCriterion.
The "Item Type" ATPortalTypeCriterion, despite its name, actually used the
"Type" index and not the "portal_type" index.

The "Type" index, indexes the Title property of a content type, and not the
portal_type.

This can cause problems when you're expecting the "Item Type" criterion to be
able to differentiate between different content types with the same Title
property (which currently happens for example with VSEvents, VSSubEvents and Events).