Allow form tabbing to be used across different DOM documents.
The existing process would select the form that wanted tabs, then
would pass the form selector into jquerytools.  This forced
jquerytools to use a $('selector') call that used the `document` from
the current frame.  If we just remove the form selector, since it's
only ever selecting on the form in the first place, and leave the pane
selector, then the form tab initialization can happen across DOM
documents such as allowing the toolbar <iframe> to initialize tabs in
a modal overlay in the parent frame.