Don't swallow exceptions on object paste
The current fix for #9365 does not fix the issue. The transaction
is still commited to the ZODB. This may lead to an inconsistent state.
The current best solution is to reraise the exception which unfortunately
leads to a HTTP 500 error.

See https://dev.plone.org/ticket/9365 for the details.

Conflicts:
	Products/CMFPlone/skins/plone_scripts/folder_paste.cpy
	Products/CMFPlone/skins/plone_scripts/object_paste.cpy
	Products/CMFPlone/tests/testCutPasteSecurity.py
	docs/CHANGES.rst