Retrieve flag var from right scope
Thanks to sentry I found this bug.
When working on objects during creation (AT) and syndication is
enabled for this specific content type, submitting the form with failures
like missing name will result in errros in the backend, triggered by
safe_unlock. Safe_unlock should not be called on submitting forms
and there is unloadForm.js that sets the submitting variable in
unlockOnFormUnload.js. Unfortunately the variable was looked up in the
wrong context.