Fix JS resource viewlet HTML syntax error
<script> tags cannot be self-closing since the content of the element is
considered to be the script if src cannot be retrieved.  It seems some
parsers, such as Diazo's handle this gracefully and add the close tags,
but others, such as Chrome, don't and when rendered the first
self-closing script tag consumes the rest of the HTML as if it were a
script until it finds a closing </script> tag.  In this case that means
the page appears completely empty.