Prevent patterns from being scanned/initialized multiple times.
For refernce, see my comment here:
https://github.com/Patternslib/Patterns/issues/396#issuecomment-73722048

This commit is also relate to https://github.com/plone/mockup/pull/462

*Here follow the same explanation as in that pull request:*

If we have multiple bundles that all pull in pat-registry and then call
registry.scan, then patterns will be scanned and initialized multiple times.

Instead, we need to call registry.init if it's not been called before. If
it's been called before, we don't need to do anything because the patterns will
automatically be registered as soon as they are loaded by require.js and
evaluated by the JS engine and then automatically scanned by the registry.

*Note:* this change relies on the Patternslib registry and won't work with
the fork @vangheem made of pat-registry. Hopefully that fork can now be
discarded.