Remove Zope branding
This is absolutely, positively **not** to be interpreted as anti-Zope sentiment, rather it's a pro-consistent-Plone-marketing statement. We no longer need to associate Plone with Zope technologies in any public-facing way **inside** the Plone product nor should we. Strategic alignment with Zope on websites, social media, etc is still absolutely fine and should still be encouraged (e.g. I still <3 Zope!) But the Plone product itself should scream "I am Plone!" not "I am a mixture of Plone and Zope!" The latter serves no purpose anymore but to confuse people and hinder Plone's mission to be the leading open source enterprise CMS.