Give better options in deprecation warning for PloneBatch.__len__.
Until now we were advocating use of the 'pagesize' attribute but we are
actually using the 'length' attribute.

I do not understand why we deprecated it in the first place, except
when this is because of possible confusion of batch length and
sequence length.  But then we should be pointing them at both options,
which we now do.

Also, adding a '__nonzero__' method in the previous commit to handle
'bool(batch)' calls, really helps in avoiding calls to the '__len__'
method.