folder_listing: avoid rebatching an empty foldercontents batch.
If 'folderContents' was a Batch but it was empty, we would batch this batch,
which works, but gives a DeprecationWarning because in plone.batching
you then start asking for len(self._sequence), which is deprecated on a PloneBatch.

This uses newer style 'A if condition else B' instead of 'if condition and A or B'.