When creating site, reapply all profiles.
Until GenericSetup 1.7.7 this is the only strategy available.
GS 1.7.8 (unreleased) offers more strategies, with as default to upgrade
already applied profiles.  Somehow we rely on reapplying all profiles.
Without this, plone.app.discussion misses its workflow settings.
And plone.api tests give a ComponentLookupError for the IMimeTypesRegistryTool.

We now override the strategy when this is possible.