Registered post_handler instead of plone-final.
The plone-final import step now does nothing.

Instead, we redefined the old handler as a post_handler explicitly for
our main profile.  This is guaranteed to really run after all other
import steps, which was never possible in the old way.

The plone-final step is kept for backwards compatibility.

This requires Products.GenericSetup 1.8.2.