Fixed html validation: section lacks heading.
Renamed some sections to divs.
These were not really sections.

See http://w3c.github.io/html/sections.html#the-section-element
which says:

The section element is not a generic container element. When an
element is needed only for styling purposes or as a convenience for
scripting, authors are encouraged to use the div element instead. A
general rule is that the section element is appropriate only if the
element’s contents would be listed explicitly in the document’s
outline.