Fixed displaying the body text of a feed item.
This is when render_body is switched on in the Syndication settings.

We were overwriting the definition of the `view` variable, which made
it fail:
AttributeError: 'str' object has no attribute 'context'

Strangely, this does not fail in the tests.

Note that we caught the error so no error was visible to a visitor.

Also note that getting the content-core macro may fail anyway, for
example for a Folder, because its default view does not have this
macro.