fix csrf.txt: use @@new-user instead of @@register.
This fixes a problem checking the csrf authenticator after removing
ZopeTestCase.  I don't know why it became a problem after that, but at
this point the site admin is logged in and tries to add a new user.
For that, we should have always used the new-user page, and not the
register page.

Also, no longer go to load the Site Setup page: we are not doing anything there.