WorkflowTool: Include Inactive Content In Worklists
WorkflowTool.getWorklistsResults uses the catalog to gather its results.
By default the catalog tool excludes inactive content from the search
results if the current user lacks the permission to view inactive
content in the current context. This can result in incomplete results,
since the method is supposed to return the work list results for the
whole site.

This commit bypasses the the catalog's inactive content filter.