Inline validation for AT: use url of form action as basis.
The base attribute gives 'context-url/' for a folder, and 'context-url/edit' for a page.
location.href gives 'context-url/edit' for both, but it may be 'edit?foo=bar'.
So the form action seems best.
That is also what the other inline validations in this file use (for formlib and z3c.form).
Issue https://github.com/plone/Products.CMFPlone/issues/1982