Fixed unstable Markup Control Panel robot test again.
The trick seems to be to first check the allowed type that we want to add,
and then uncheck the other ones.
Otherwise you get a slightly belated validation error before pressing Save,
and the Save has no effect.

(If this does not help, we really have to add a 'Sleep  1' before pressing Save.)