Depend on new package Products.isurlinportal.
This contains the `isURLInPortal` method that was split off from our `URLTool`.
See https://github.com/plone/Products.CMFPlone/issues/3150.

We depend on a version less than 2.0.0.
This is in the expectancy that such a breaking version will be released at some point,
probably when dropping support for Python 2 or for Zope 2.