Removed dependency on ``Products.TemporaryFolder``.
See https://github.com/plone/Products.CMFPlone/issues/2957.

Note: in your `plone.recipe.zope2instance` buildout part, you must set `zodb-temporary-storage = off`,
otherwise you get errors when starting Plone.
We could make this the default in the recipe, but that can only be done in a branch for Plone 6.
It feels too early to make such a branch, as much can still be improved to the recipe on Plone 5.2, leading to double efforts.