Addon list for upgrade: try to upgrade default profile of plone.app.iterate.
This package at the moment still has a profile called plone.app.iterate.
In 3.1.5 (used since Plone 5.0.3), the default profile was added.
The old name is kept for compatibility until now, but should be removed.
Perhaps this may need an upgrade step in plone.app.iterate for people who used the old name.

But with the old name in the addon list, `plone.app.iterate:default` was never upgraded automatically.