On Zope root, create Volto site by default.
When the `plone.volto` package is not available, the buttons are the same as before,
and will create a Classic Plone site.

When `plone.volto` is there, you get an extra button for Classic:

- 'Create a new Plone site' will create a Volto site.
- 'Create Classic Plone site' will create a Classic site.
- Advanced still gives you all options.

When installing Volto, we install the caching and Barcelonata profiles just like a Classic site,
although we could choose to not install Barceloneta.
As extra, we install `plone.volto:default` and `plone.volto:default-homepage`.
(Technically, `default-homepage` would already pull in `default`, but let's do this explicitly.)