feat(upgrade): Also upgrade Volto and REST API
Now that Volto is or will be a supported optional core Plone add-on, it's profiles
should also be upgraded when the Plone portal is upgraded if it has been
installed/imported.

There may be other optional add-ons that have come to be supported by Plone core itself
since this list was last updated.  An audit should be done for any such add-ons and they
should be added as well.

WIP: These changes are intended to demonstrate this approach.  It can be used as the
basis for the real changes, but the impact of the changed should be reviewed by someone
familiar with the reasons for the original approach, tests should be added and resulting
issues resolved.

Refs #3453