Add GitHub action to test PR on coredev.
This is a gitHub Action for testing a package PR on buildout.coredev.
For now (May 2022) it is an experiment, but I think we need it.
Reasons:

1. The robot tests on Jenkins are flaky/unstable, which most likely is due to several robot jobs running parallel on one node.
2. Python 3.10 is not working with the 'Python Shining Pandas' plugin we use on Jenkins.  We had it running with a different script for a while, but these changes got lost.  Should be restorable, but let's try GHA.

One thing to watch out for, is that robot tests are not always reported as failures.
On the one hand there is config for that on Jenkins:
https://github.com/plone/jenkins.plone.org/issues/297
On the other hand I see the same problem locally:
https://github.com/plone/Products.CMFPlone/issues/3537
But those two problems seem solved by now (June 2022).

We use a reusable workflow in [`plone/.github`](https://github.com/plone/.github/blob/main/reusable-workflows/coredev.yml).

Aftwerwards, we could install it in all relevant Plone repos using https://github.com/asottile/all-repos
But nothing too hasty. :-)