Make safe html filter robot tests more robust.
Unfortunately, it seems we need to sleep a bit for this.
The problem I saw is that the rich text we set is not actually saved.
A harmless paragraph tag did not end up on the resulting page, or at least not always.
Adding some extra lines in document.pt showed that the text attribute was actually empty in this case.

Locally I now ran the `test_controlpanel_filter.robot` tests ten times and they passed.