Restore imageScales / image_scales in patterns settings.
Otherwise the new code with picture variants is only used when you have the right mockup branch,
and mockup running, and the resource registries setup to use it,
or the mockup changes should have landed in plone.staticresources already.

Note that the breakage is not too bad though:
if imageScales is not in the patterns settings and you still have TinyMCE from old staticresources,
then TinyMCE will still work.  It will just use an internal copy of the imageScales,
which for example is missing the new greater and huge scales.
With the current commit, we support both old and new TinyMCE, as long as we think this is needed.