Allow bundles to be rendered after "theme" and "custom" bundles.
Bundles can now depend on the automatically added "theme" and "custom" bundles
from plone.app.theming and be rendered after those (e.g. after the
plonetheme.barceloneta CSS).

This allows to override a theme with custom CSS from a bundle instead of having
to add the CSS customizations to the registry via the "custom_css" settings.
As a consequence, theme customization can now be done in the filesystem in
ordinary CSS files instead of being bound to a time consuming workflow which
involces upgrading the registry after every change.