Remove deferred import for `INavigationRoot`.
We said we would remove this in Plone 5.0 because it was moved to `plone.app.layout`.  Meanwhile it has been moved to `plone.base`.

Should be fine to remove: a Plone Site that still has this location as a persistent marker interface, needs to be migrated to Plone 5.2 and then go through `zodbupdate` for Python 3, so the persistent interface location will be updated.