RegistrationTool: add method principal_id_or_login_name_exists.
This is factored out from the `isMemberIdAllowed` method, which now calls this after checking the allowed member id pattern.

Main use case: let `plone.app.users` call the new method when saving personal information.
This is done when email-as-login is used.  In that case, we currently check if the email address would be valid as id, and if there is not yet a user or group with this id or login name.
Problem is that some valid emails are not valid as user id, for example an email address with a quote, and this makes the check fail.