Remove test for adding an English site with a Dutch browser.
This test was skipped if plone.distribution was available.
Now we are testing it on GHA without plone.distribution.
And it fails because the new site has no text at all.
This is more or less on purpose, as the standard/fallback addsite view is stripped down.
Try 'bin/instance-cmfplone fg' in coredev 6.2, create a Plone Site, and you see there is no text.