Automatically set the ajax_load request parameter. (#4169)
* Automatically set the ajax_load request parameter.

Most JavaScript libraries send the "HTTP_X_REQUESTED_WITH" request header set
to "XMLHttpRequest" for AJAX requests. For AJAX requests we set the "ajax_load"
parameter on the request object before traversing, if it was not already set.

This is further used to switch to the ajax_main_template and to turn off Diazo
transformations.

For AJAX requests which need a full-page rendering this automatic setting can
be turned off by adding `?ajax_load=0` as query string to the URL.

* Automatically set the ajax_load request parameter.

Set ajax_load parameter even earlier at the ZPublisher.interfaces.IPubStart event.

* Automatically set the ajax_load request parameter.

Get "plone.ajax_marker" registry parameter to enable/disable automatic ajax marking.
The "ajax_marker" configuration option is defined in
"plone.base.interfaces.controlpanel.ISiteSchema" and defaults to "False", so no
automatic setting.

* use ZPublisher.interfaces.IPubAfterTraversal" event, where the registry can be retrieved but its already too late in the chain and the non-ajax main template was alreday instantiated.

* Automatically use the AJAX main template.

Use the ajax main template if:

- ajax_load translate to True
- or if we have an xhr request and plone.use_ajax (TBD) is set to True.

Don't use the ajax main template if ajax_load is set to a falsy value or we're
not in an xhr request.

* Cleanup and remove set_ajax event subscriber.

* fixup! use plone_layout utils.

* fixup! tst fixes

* fixup! pre-commit

* PR review: Adapt to changed use_ajax_main_template option from plone.base.

* use_ajax now a property.

* Move use_ajax from plone.app.layout.global.layout PR to here.

* USe is_ajax.

* Test fixes.

* fixup

* Register the IClassicUISchema registry entries from plone.base.